Rename /constants/ subpackages to /constant/ for uniformity
