+++ api/src/main/java/com/example/echo_api/modules/clerk/controller/ClerkDevController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.example.echo_api.modules.clerk.service.ClerkDevService;
 import com.example.echo_api.modules.user.entity.User;
-import com.example.echo_api.shared.constants.ApiRoutes;
+import com.example.echo_api.shared.constant.ApiRoutes;
 
 import lombok.RequiredArgsConstructor;
