+++ api/src/main/java/com/example/echo_api/modules/post/controller/PostInteractionController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.echo_api.modules.post.service.PostInteractionService;
-import com.example.echo_api.shared.constants.ApiRoutes;
+import com.example.echo_api.shared.constant.ApiRoutes;
 
 import lombok.RequiredArgsConstructor;
