+++ api/src/main/java/com/example/echo_api/modules/profile/controller/ProfileViewController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.example.echo_api.modules.profile.dto.response.ProfileDTO;
 import com.example.echo_api.modules.profile.dto.response.SimplifiedProfileDTO;
 import com.example.echo_api.modules.profile.service.ProfileViewService;
-import com.example.echo_api.shared.constants.ApiRoutes;
+import com.example.echo_api.shared.constant.ApiRoutes;
 import com.example.echo_api.shared.pagination.OffsetLimitRequest;
 import com.example.echo_api.shared.pagination.PageDTO;
 import com.example.echo_api.shared.validation.annotations.Limit;
