+++ api/src/main/java/com/example/echo_api/security/ClerkOnboardingFilter.java
@@ -12,7 +12,7 @@ import org.springframework.web.filter.OncePerRequestFilter;
 
 import com.example.echo_api.config.ClerkConfig;
 import com.example.echo_api.config.ErrorMessageConfig;
-import com.example.echo_api.shared.constants.ApiRoutes;
+import com.example.echo_api.shared.constant.ApiRoutes;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
