+++ api/src/test/java/com/example/echo_api/modules/clerk/controller/ClerkControllerTest.java
@@ -22,7 +22,7 @@ import com.example.echo_api.exception.custom.unauthorised.WebhookVerificationExc
 import com.example.echo_api.modules.clerk.service.ClerkOnboardingService;
 import com.example.echo_api.modules.clerk.service.ClerkWebhookService;
 import com.example.echo_api.modules.user.entity.User;
-import com.example.echo_api.shared.constants.ApiRoutes;
+import com.example.echo_api.shared.constant.ApiRoutes;
 import com.example.echo_api.shared.dto.ErrorDTO;
 
 /**
