+++ api/src/test/java/com/example/echo_api/modules/post/controller/PostInteractionControllerTest.java
@@ -17,7 +17,7 @@ import com.example.echo_api.config.ErrorMessageConfig;
 import com.example.echo_api.exception.custom.conflict.AlreadyLikedException;
 import com.example.echo_api.exception.custom.notfound.ResourceNotFoundException;
 import com.example.echo_api.modules.post.service.PostInteractionService;
-import com.example.echo_api.shared.constants.ApiRoutes;
+import com.example.echo_api.shared.constant.ApiRoutes;
 import com.example.echo_api.shared.dto.ErrorDTO;
 
 /**
