+++ api/src/test/java/com/example/echo_api/modules/post/controller/PostManagementControllerTest.java
@@ -24,7 +24,7 @@ import com.example.echo_api.exception.custom.badrequest.InvalidParentIdException
 import com.example.echo_api.exception.custom.forbidden.ResourceOwnershipException;
 import com.example.echo_api.modules.post.dto.request.CreatePostDTO;
 import com.example.echo_api.modules.post.service.PostManagementService;
-import com.example.echo_api.shared.constants.ApiRoutes;
+import com.example.echo_api.shared.constant.ApiRoutes;
 import com.example.echo_api.shared.dto.ErrorDTO;
 import com.fasterxml.jackson.databind.ObjectMapper;
